Navigation: use shared default implementation of fetchLinkSuggestions on post, site and navigation editor
