+++ packages/block-editor/src/index.js
@@ -9,5 +9,6 @@ import '@wordpress/rich-text';
 import './hooks';
 export * from './components';
 export * from './utils';
+export * from './settings';
 export { storeConfig, store } from './store';
 export { SETTINGS_DEFAULTS } from './store/defaults';
