+++ packages/block-editor/src/settings/defaults/__experimental-fetch-link-suggestions/index.js
@@ -1,54 +1,37 @@
 /**
+ * External dependencies
+ */
+import { map, flatten } from 'lodash';
+
+/**
  * WordPress dependencies
  */
 import apiFetch from '@wordpress/api-fetch';
 import { addQueryArgs } from '@wordpress/url';
-/**
- * External dependencies
- */
-import { flatten, map } from 'lodash';
 import { decodeEntities } from '@wordpress/html-entities';
 import { __ } from '@wordpress/i18n';
 
 /**
- * Fetches link suggestions from the API. This function is an exact copy of a function found at:
- *
- * packages/editor/src/components/provider/index.js
- *
- * It seems like there is no suitable package to import this from. Ideally it would be either part of core-data.
- * Until we refactor it, just copying the code is the simplest solution.
+ * Fetches link suggestions from the API.
  *
  * @param {string} search
  * @param {Object} [searchArguments]
  * @param {number} [searchArguments.isInitialSuggestions]
  * @param {number} [searchArguments.type]
  * @param {number} [searchArguments.subtype]
+ * @param {number} [searchArguments.page]
  * @param {Object} [editorSettings]
  * @param {boolean} [editorSettings.disablePostFormats=false]
  * @return {Promise<Object[]>} List of suggestions
  */
 
-export default function fetchLinkSuggestions(
+const fetchLinkSuggestions = async (
 	search,
-	{ isInitialSuggestions, type, subtype } = {},
+	{ isInitialSuggestions, type, subtype, page, perPage: perPageArg } = {},
 	{ disablePostFormats = false } = {}
-) {
-	const perPage = isInitialSuggestions ? 3 : 20;
+) => {
+	const perPage = perPageArg || isInitialSuggestions ? 3 : 20;
 
-	const linkTypes = [ 'post', 'term', 'post-format' ];
-
-	linkTypes.forEach( ( linkType ) => {
-		if ( ! type || type === linkType ) {
-			apiFetch( {
-				path: addQueryArgs( '/wp/v2/search', {
-					search,
-					per_page: perPage,
-					type: 'post',
-					subtype,
-				} ),
-			} ).catch( () => [] ); // fail by returning no results
-		}
-	} );
 	const queries = [];
 
 	if ( ! type || type === 'post' ) {
@@ -56,6 +39,7 @@ export default function fetchLinkSuggestions(
 			apiFetch( {
 				path: addQueryArgs( '/wp/v2/search', {
 					search,
+					page,
 					per_page: perPage,
 					type: 'post',
 					subtype,
@@ -69,6 +53,7 @@ export default function fetchLinkSuggestions(
 			apiFetch( {
 				path: addQueryArgs( '/wp/v2/search', {
 					search,
+					page,
 					per_page: perPage,
 					type: 'term',
 					subtype,
@@ -82,6 +67,7 @@ export default function fetchLinkSuggestions(
 			apiFetch( {
 				path: addQueryArgs( '/wp/v2/search', {
 					search,
+					page,
 					per_page: perPage,
 					type: 'post-format',
 					subtype,
@@ -91,11 +77,18 @@ export default function fetchLinkSuggestions(
 	}
 
 	return Promise.all( queries ).then( ( results ) => {
-		return map( flatten( results ).slice( 0, perPage ), ( result ) => ( {
-			id: result.id,
-			url: result.url,
-			title: decodeEntities( result.title ) || __( '(no title)' ),
-			type: result.subtype || result.type,
-		} ) );
+		return map(
+			flatten( results )
+				.filter( ( result ) => !! result.id )
+				.slice( 0, perPage ),
+			( result ) => ( {
+				id: result.id,
+				url: result.url,
+				title: decodeEntities( result.title ) || __( '(no title)' ),
+				type: result.subtype || result.type,
+			} )
+		);
 	} );
-}
+};
+
+export default fetchLinkSuggestions;
