+++ packages/block-editor/src/settings/index.js
@@ -1,0 +1,1 @@
+export { default as __experimentalFetchLinkSuggestions } from './defaults/__experimental-fetch-link-suggestions';
