+++ packages/edit-site/src/index.js
@@ -1,14 +1,12 @@
 /**
  * WordPress dependencies
  */
-import apiFetch from '@wordpress/api-fetch';
-import { addQueryArgs } from '@wordpress/url';
-import { __ } from '@wordpress/i18n';
 import {
 	registerCoreBlocks,
 	__experimentalRegisterExperimentalCoreBlocks,
 } from '@wordpress/block-library';
 import { render } from '@wordpress/element';
+import { __experimentalFetchLinkSuggestions as fetchLinkSuggestions } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -18,32 +16,6 @@ import './hooks';
 import './store';
 import Editor from './components/editor';
 
-const fetchLinkSuggestions = ( search, { perPage = 20 } = {} ) =>
-	apiFetch( {
-		path: addQueryArgs( '/wp/v2/search', {
-			per_page: perPage,
-			search,
-			type: 'post',
-			subtype: 'post',
-		} ),
-	} )
-		.then( ( posts ) =>
-			Promise.all(
-				posts.map( ( post ) =>
-					apiFetch( { url: post._links.self[ 0 ].href } )
-				)
-			)
-		)
-		.then( ( posts ) =>
-			posts.map( ( post ) => ( {
-				url: post.link,
-				type: post.type,
-				id: post.id,
-				slug: post.slug,
-				title: post.title.rendered || __( '(no title)' ),
-			} ) )
-		);
-
 /**
  * Initializes the site editor screen.
  *
@@ -51,7 +23,10 @@ const fetchLinkSuggestions = ( search, { perPage = 20 } = {} ) =>
  * @param {Object} settings Editor settings.
  */
 export function initialize( id, settings ) {
-	settings.__experimentalFetchLinkSuggestions = fetchLinkSuggestions;
+	settings.__experimentalFetchLinkSuggestions = (
+		searchText,
+		searchOptions
+	) => fetchLinkSuggestions( searchText, searchOptions, settings );
 	settings.__experimentalSpotlightEntityBlocks = [ 'core/template-part' ];
 
 	registerCoreBlocks();
